feat(benchmark): moved from utils to test_utils
